+++ TriviaApp/app/src/main/java/com/example/triviaapp/viewModels/Uis/Uis.kt
@@ -6,7 +6,7 @@ import com.example.triviaapp.componentes.Tarjeta
 data class PaginaAjustesUi(
     val nombreTriv:String="",
     val categoria: String="1",
-    val preguntas: Float=0f
+    val preguntas: Int=0
     )
 
 data class ElegirRespUIState(
